haskell: Add "tdd" wrapper around `ghcid`
This adds a small helper script `tdd` which works around some of the
issues with `ghcid` where it doesn't handle very gracefully
recompilation of the `library` when using it as a test runner. Changes
to the tests are instantly applied and re-run, but library changes
aren't detected.

This isn't pretty, but the best I've got so far ...

See: https://stackoverflow.com/a/74276514/171298